+++ src/components/Dropdown/index.ts
@@ -1,3 +1,4 @@
 import Dropdown from '~/components/Dropdown/Dropdown';
+import DropdownExample from '~/components/Dropdown/DropdownExample';
 
-export { Dropdown };
+export { Dropdown, DropdownExample };
